pyproject.toml
[project]
name = "finalproj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepagents>=0.1.4",
    "dotenv>=0.9.9",
    "fastapi>=0.120.0",
    "langchain[google-genai]>=1.0.2",
    "pydantic-settings>=2.11.0",
    "tavily-python>=0.7.12",
    "uvicorn>=0.38.0",
]
